requirements.txt
numpy==2.2.2
jmetalpy==1.7.0